refactor: refactor code and update configuration files
