+++ .husky/pre-commit
@@ -1,1 +1,3 @@
+npm run lint
 npx lint-staged
+npm run test:coverage
