+++ package.json
@@ -4,16 +4,19 @@
   "description": "This repository provides a Node.js application for demonstrating JWT authentication and authorization. It includes features such as user registration, user login, and authorization process. The application is built using Express.js, JWT for authentication, Sequelize ORM for database interactions, and MySQL for database storage.",
   "main": "app/server.js",
   "type": "module",
+  "engines": {
+    "node": ">=20.6.0"
+  },
   "scripts": {
     "start": "node --env-file .env app/server.js ",
     "dev": "nodemon --env-file .env app/server.js",
-    "sync-db": "node custom_migration/sync-db.js",
-    "create-table": "sequelize-cli migration:generate --name create_table",
+    "sync-db": "node database/custom_migration/sync-db.js",
+    "create-table": "sequelize-cli migration:generate --name",
     "migrate": "sequelize-cli db:migrate",
     "migrate:undo": "sequelize-cli db:migrate:undo",
     "migrate:undo:all": "sequelize-cli db:migrate:undo:all",
     "seed": "sequelize-cli db:seed:all",
-    "generate:seed": "sequelize-cli seed:generate --name demo-user",
+    "generate:seed": "sequelize-cli seed:generate --name",
     "lint:clear": "rm -rf ~/.eslintcache",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
@@ -25,6 +28,14 @@
     "updateSnapshots": "jest -u --coverage=false"
   },
   "jest": {
+    "coverageThreshold": {
+      "global": {
+        "branches": 80,
+        "functions": 80,
+        "lines": 80,
+        "statements": 80
+      }
+    },
     "collectCoverage": true,
     "collectCoverageFrom": [
       "app/**/*.js",
